+++ core/engine/src/builtins/string/mod.rs
@@ -694,15 +694,16 @@ impl String {
         // 3. Let n be ? ToIntegerOrInfinity(count).
         match args.get_or_undefined(0).to_integer_or_infinity(context)? {
             IntegerOrInfinity::Integer(n)
-                if u64::try_from(n)
+                if usize::try_from(n)
                     .ok()
-                    .and_then(|n| n.checked_mul(len as u64))
-                    .is_some_and(|total_len| total_len <= (Self::MAX_STRING_LENGTH as u64)) =>
+                    .and_then(|n| n.checked_mul(len))
+                    .is_some_and(|total_len| total_len <= Self::MAX_STRING_LENGTH) =>
             {
                 if string.is_empty() {
                     return Ok(js_string!().into());
                 }
-                let n = n as usize;
+
+                let n = n as usize; // safe due to guard above
                 let mut result = Vec::with_capacity(n);
 
                 std::iter::repeat_n(string.as_str(), n).for_each(|s| result.push(s));
